Temporarily revert loading dependencies
This reverts the recent change
`Update settings to load fat jar but not analyse dependencies`
so that we can properly investigate the effect of this change on
dashboard. Currently we cannot access the detailed JaCoCo logs because
the VMs from the time this change was originally made no longer exist.
After a dashboard run has been triggered with the reverted change, the
change should immediately be restored in a follow-up PR.